+++ src/javascripts/security/role/full.ctrl.js
@@ -1,9 +1,9 @@
-import roleApi from '../common/roleApi.service';
+import roleApi from '../../common/roleApi.service';
 import uiNotification from 'angular-ui-notification';
-import prevousState from '../common/previousState.service.js';
-import authorizationApi from '../authentication/authorizationApi.service';
-import schemaService from '../common/schema.service';
-import jsonEdit from '../common/jsonEdit/jsonEdit.directive';
+import prevousState from '../../common/previousState.service.js';
+import authorizationApi from '../../authentication/authorizationApi.service';
+import schemaService from '../../common/schema.service';
+import jsonEdit from '../../common/jsonEdit/jsonEdit.directive';
 
 let ctrlName = 'RoleFullCtrl';
